[REFACTOR]: Change endpoint name
