+++ constants/httpManagement.js
